add testbench
this has already been in the making for a long while at this point and i
just need to get a version committed at some point.

we're already well past the 5th printed version (might already be at 10,
depending on what counts and what doesn't), 10s of work hours and kilos
of pla in.

it's very much still in development but i want to remove the commented
out old stuff but i don't want to lose the historic record